Rearranged files acc to file type
