+++ SECURITY.md
@@ -1,3 +1,4 @@
 # Security Policy
 
-Since this repo mainly deals with DS and Algos, if you've encountered a security issue, you should probably wear a tinfoil hat :)
+Since this repo mainly deals with DS and Algos, if you've encountered a security issue,
+you should probably wear a tinfoil hat :)
